Rename document.md to README.md
The document.md file was renamed to README.md to better adhere to Github's standard rendering of top-level project documentation files. This change was made to enhance the visibility and easy accessibility of the instructions or explanations about the code in the repository for a user or new collaborator browsing the repository.